move v2ext to conditionHelpers
